# Fix unaligned access in ACLE based crc32