This fixes a rightful complaint from the alignment sanitizer that we alias memory in an unaligned fashion. A nice added bonus is that this improves performance a tiny by on the larger buffers, perhaps due to loops that idiomatically decrement a count and increment a single buffer pointer rather than the maze of conditional pointer reassignments.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **Bug Fixes**
  - Improved buffer handling for CRC computation, ensuring accurate processing of data.
  - Enhanced testing for alignment issues in CRC32 calculations.

- **New Features**
  - Streamlined the CRC calculation process by utilizing local variables for buffer data, enhancing performance.
  - Introduced a new test class to verify CRC32 functionality with varying alignment offsets.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->