@ccawley2011 2024-12-21
The code is supposed to ensure that the pointers are aligned by performing smaller loads until alignment is reached, so if that isn't happening it's likely that there's something else wrong.

Using `memcpy` will likely hurt performance on operating systems where unaligned access is forbidden, since it'll need to load each word one byte at a time before combining them.