@KungFuJesus 2024-12-21
> The code is supposed to ensure that the pointers are aligned by performing smaller loads until alignment is reached, so if that isn't happening it's likely that there's something else wrong.
> 
> Using `memcpy` will likely hurt performance on operating systems where unaligned access is forbidden, since it'll need to load each word one byte at a time before combining them.

It doesn't though, as the compiler optimizes memcpys of fixed lengths to call builtins that translate the calls to direct unaligned loads on architectures that support it (which ARM does).  This is more to ensure that we're calling well defined behavior within C.  The sanitizer detected unaligned access at runtime, so the ladder of ifs had a lapse in proper alignment checks.  Actually I think I see where the alignment checks are failing because the & trick for powers of 2 requires a subtraction of one.  This is being done for the first one but not subsequent ones.