@KungFuJesus 2024-12-21
Now the access should actually be aligned to the data type being read (for the slight benefit of speed on any particularly ARM chip).