@ccawley2011 2024-12-21
> It doesn't though, as the compiler optimizes memcpys of fixed lengths to call builtins that translate the calls to direct unaligned loads on architectures that support it (which ARM does).

The behaviour on ARMv6 and newer is dependent on how the operating system configures alignment exceptions - if they're enabled (or ARMv4/v5 behaviour is enabled on ARMv6), applications and libraries will need to be compiled without unaligned access, and the behaviour observed in PR #1548 will occur instead. RISC OS is an example of a platform that both supports ARMv8 with CRC extensions and has alignment exceptions enabled by default.