@KungFuJesus 2024-12-21
Hmm, so it sounds like the may_alias attribute is needed for those cases.  I can write this so that it just dereferences a pointer but it's hard to imagine a case where the compiler wouldn't treat that the same way as us calling a fixed sized memcpy as it sees we're assigning the pointer from something that only has 1 byte alignment.  FWIW, with the full test corpora this passes without memcpy and simply just deref'ing the pointers as the respective types, so you shouldn't be seeing unaligned access regardless.  The memcpys are more there as a guard rail (though it sounds like one you can't rely on if you're on RISC OS).