@KungFuJesus 2024-12-21
@ccawley2011 It's now memcpy free.  We'll have to rely on the alignment sanitizer to police it here.