@Dead2 2024-12-22
### Aarch64 rpi5
#### Before PR
```
crc32/acle/1              2.50 ns         2.50 ns    279907009
crc32/acle/8              4.17 ns         4.17 ns    167944184
crc32/acle/12             4.17 ns         4.17 ns    167945907
crc32/acle/16             4.58 ns         4.58 ns    152674019
crc32/acle/32             6.67 ns         6.67 ns    104957581
crc32/acle/64             8.34 ns         8.34 ns     83971953
crc32/acle/512            35.1 ns         35.1 ns     19920130
crc32/acle/4096            226 ns          226 ns      3092795
crc32/acle/32768          1720 ns         1720 ns       406891
crc32/acle/262144        14683 ns        14683 ns        47675
crc32/acle/4194304      240527 ns       240526 ns         2905
```
#### After PR
```
crc32/acle/1              2.50 ns         2.50 ns    279910077
crc32/acle/8              4.17 ns         4.17 ns    167945750
crc32/acle/12             4.17 ns         4.17 ns    167946288
crc32/acle/16             4.58 ns         4.58 ns    152674642
crc32/acle/32             6.67 ns         6.67 ns    104964779
crc32/acle/64             8.34 ns         8.34 ns     83969360
crc32/acle/512            35.7 ns         35.7 ns     19554344
crc32/acle/4096            226 ns          226 ns      3092795
crc32/acle/32768          1720 ns         1720 ns       406898
crc32/acle/262144        14666 ns        14665 ns        47730
crc32/acle/4194304      240605 ns       240604 ns         2906
```
Performance here seems to be pretty much identical.