@phprus 2024-12-22
Unaligned access is not fixed

Ubuntu GCC AARCH64 No NEON UBSAN (https://github.com/zlib-ng/zlib-ng/actions/runs/12447135765/job/34750296246?pr=1833):

```
/home/runner/work/zlib-ng/zlib-ng/arch/arm/crc32_acle.c:40:18: runtime error: load of misaligned address 0x00550046cc52 for type 'uint32_t', which requires 4 byte alignment
```