@KungFuJesus 2024-12-22
Ugh, I ran through the whole data test corpus without issue but alignment is just kind of a chance thing if we can't generate unaligned input. Maybe I'll need to test it that way, forcing unaligned input.