@KungFuJesus 2024-12-22
> Test for unaligned access (pseudocode):
> 
> ```c
> char buf[32];
> size_t len = sizeof(buf);
> for (size_t i = 0; i < len; i++)
> {
>     buf[i] = 'a' + (i % 26);
> }
> 
> for (size_t offset = 0; offset < sizeof(uint64_t); offset++)
> {
>     for (size_t current_len = 1; current_len < (current_len - offset); current_len++)
>     {
>         c = crc(0, buf + offset, current_len);
>         ignore(c);
>     }
> }
> ```

Yep that's more or less how I tested it, I just did it in the benchmark code where we start with a 64 byte aligned address and walked values 1-7 for that.