@phprus 2024-12-22
> Yep that's more or less how I tested it, I just did it in the benchmark code where we start with a 64 byte aligned address and walked values 1-7 for that.

Maybe add this test to zlib-ng?