@KungFuJesus 2024-12-22
> > Yep that's more or less how I tested it, I just did it in the benchmark code where we start with a 64 byte aligned address and walked values 1-7 for that.
> 
> Maybe add this test to zlib-ng?

I mean my test consisted of running that in conjunction with ubsan.  If not compiled with ubsan it's not much of a test but I suppose I could write it as at least one runner will run it with that.  I've never written gtests before, though.