@KungFuJesus 2024-12-22
Ok, I added a gtest specifically for this.